修改爬虫逻辑

+++ src/main/java/com/tingnichui/pojo/po/DailyRecord.java
@@ -65,7 +65,7 @@ public class DailyRecord {
      * 成交额(万)
      */
     @TableField(value = "amount")
-    private Long amount;
+    private BigDecimal amount;
 
     /**
      * 涨跌比率
